# openapi3.yaml
openapi: 3.0.1
info:
  title: lookup-tables
  description: Service which serve secret list data like countries, clasiification, etc.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /country:
    get:
      operationId: getCountryList
      tags:
        - getCountryList
      summary: get country list
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/countryList'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /country/excludeFields:
    post:
      operationId: getCountryListExcludeFields
      tags:
        - GetCountryListExcludeFields
      summary: get country list and exclude fields
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items: 
                $ref: '#/components/schemas/countryField'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/countryList'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /classification:
    get:
      operationId: getClassificationList
      tags:
        - getClassificationList
      summary: get classification list
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/classificationList'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
components:
  schemas:
    country:
      type: object
      required:
        - code
      properties:
        code:
          type: number
          format: int64
        name:
          type: string
    countryList:
      type: array
      items:
        $ref: '#/components/schemas/country'
    countryField:
        type: string
        enum: [code, name]
    classification:
      type: object
      required:
        - id
      properties:
        id:
          type: number
          format: int64
        name:
          type: string
    classificationList:
      type: array
      items:
        $ref: '#/components/schemas/classification'
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string